+++ src/components/FAQSection/index.ts
@@ -2,8 +2,5 @@ export { default } from "./FAQSection";
 export { default as FAQSidebar } from "./FAQSidebar";
 export { default as FAQSearch } from "./FAQSearch";
 export { default as FAQContentFilter } from "./FAQContentFilter";
-export {
-  default as FAQPageContext,
-  useFAQTemplate,
-} from "./FAQSectionsContext";
-export type { FAQSection } from "./FAQSectionsContext";
+export { default as FAQPageContext, useFAQTemplate } from "./FAQPageContext";
+export type { FAQSection } from "./FAQPageContext";
